♻️ refactor(arch): move delayed-jobs interfaces to application layer
The I*DelayedJobs interfaces referenced concrete delayed-job classes in
application/jobs/, so keeping them under domain/jobs/ made the domain layer
depend on the application layer (hexagonal violation caught by the arch
domain-purity rule). Move them into application/jobs/ next to the jobs they
describe, for commands, deployments and git.
