+++ packages/commands/src/domain/jobs/index.ts
@@ -1,2 +1,1 @@
 export * from './DeployCommands';
-export * from './ICommandsDelayedJobs';
