+++ packages/deployments/src/application/jobs/IDeploymentsDelayedJobs.ts
@@ -1,4 +1,4 @@
-import { PublishArtifactsDelayedJob } from '../../application/jobs/PublishArtifactsDelayedJob';
+import { PublishArtifactsDelayedJob } from './PublishArtifactsDelayedJob';
 
 export interface IDeploymentsDelayedJobs {
   publishArtifactsDelayedJob: PublishArtifactsDelayedJob;
