+++ packages/deployments/src/domain/jobs/index.ts
@@ -1,2 +1,1 @@
 export * from './PublishArtifactsJob';
-export * from './IDeploymentsDelayedJobs';
