+++ packages/git/src/domain/jobs/index.ts
@@ -2,4 +2,3 @@ export {
   FetchFileContentInput,
   FetchFileContentOutput,
 } from './FetchFileContent';
-export { IGitDelayedJobs } from './IGitDelayedJobs';
